// frontend/src/routes/types.ts
export type StackParamList = {
  Home: undefined;
  Login: undefined;
  Cadastro: undefined;
  DashboardVagas: undefined;
  OngAcoes: undefined;
  EditarVaga: { vaga: any };
  OrganizationProfile: undefined;
  UserProfile: undefined;
  Chat: { vagaId: number; vagaNome: string };
  VagasVoluntariadas: undefined;
  MinhasVagas: undefined; 
  OngReports: undefined;
  CreateVaga: undefined;
  GerenciarInscritos: { vagaId: number; vagaNome: string };
};